requirements.txt
flask
TA-Lib
pandas
yfinance
python-binance